+++ .github/workflows/build.yaml
@@ -1,0 +1,50 @@
+name: build
+
+# Trigger the workflow on pushes to any branch
+on:
+  push:
+    branches:
+      - '**'
+
+jobs:
+  build-push:
+    runs-on: ubuntu-latest
+
+#    environment:
+#      name: ${{ github.ref == 'refs/heads/master' && 'prod' || 'dev' }}
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: setup python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.11'
+
+      - name: setup-node
+        uses: actions/setup-node@v4
+        with:
+          node-version: '16'
+
+      - name: setup-buildx
+        uses: docker/setup-buildx-action@v3
+
+#      - name: Configure AWS credentials
+#        uses: aws-actions/configure-aws-credentials@v2
+
+      - name: setup-cdk
+        run: npm install -g aws-cdk
+
+      - name: setup-venv
+        run: |
+          python -m venv ./.venv
+          source ./.venv/bin/activate
+          pip install --upgrade pip
+          pip install -r ./requirements.txt
+          
+      - name: cdk-synth
+        run: |
+          source ./.venv/bin/activate
+          cdk synth
+
+
