+++ .gitignore
@@ -1,0 +1,13 @@
+.venv/
+.idea/
+
+*.swp
+package-lock.json
+__pycache__
+.pytest_cache
+*.egg-info
+
+# CDK asset staging directory
+.cdk.staging
+cdk.out
+
