+++ nbconvert/requirements.txt
@@ -1,0 +1,35 @@
+attrs==24.2.0
+beautifulsoup4==4.12.3
+bleach==6.1.0
+certifi==2024.8.30
+charset-normalizer==3.3.2
+defusedxml==0.7.1
+fastjsonschema==2.20.0
+idna==3.8
+Jinja2==3.1.4
+jsonschema==4.23.0
+jsonschema-specifications==2023.12.1
+jupyter_client==8.6.2
+jupyter_core==5.7.2
+jupyterlab_pygments==0.3.0
+MarkupSafe==2.1.5
+mistune==3.0.2
+nbclient==0.10.0
+nbconvert==7.16.4
+nbformat==5.10.4
+packaging==24.1
+pandocfilters==1.5.1
+platformdirs==4.3.2
+Pygments==2.18.0
+python-dateutil==2.9.0.post0
+pyzmq==26.2.0
+referencing==0.35.1
+requests==2.32.3
+rpds-py==0.20.0
+six==1.16.0
+soupsieve==2.6
+tinycss2==1.3.0
+tornado==6.4.1
+traitlets==5.14.3
+urllib3==2.2.2
+webencodings==0.5.1
