+++ requirements.txt
@@ -1,0 +1,2 @@
+aws-cdk-lib==2.143.0
+constructs>=10.0.0,<11.0.0
